+++ UltitemsCyan/Items/Tier3/ViralSmog.cs
@@ -4,12 +4,14 @@ using BepInEx.Configuration;
 
 namespace UltitemsCyan.Items.Tier3
 {
+    // Notes:
+    // Luminous Shot's counting status isn't a cooldown buff so it counts for smog
 
     // TODO: check if Item classes needs to be public
     public class ViralSmog : ItemBase
     {
         public static ItemDef item;
-        private const float speedPerStackStatus = 25f;
+        private const float speedPerStackStatus = 20f;
 
         public override void Init(ConfigFile configs)
         {
@@ -22,7 +24,7 @@ namespace UltitemsCyan.Items.Tier3
                 "VIRALSMOG",
                 itemName,
                 "Increase speed per unique status effect.",
-                "Increases <style=cIsUtility>movement speed</style> by <style=cIsUtility>25%</style> <style=cStack>(+25% per stack)</style> per <style=cIsDamage>unique status</style> you have.",
+                "Increases <style=cIsUtility>movement speed</style> by <style=cIsUtility>20%</style> <style=cStack>(+20% per stack)</style> per <style=cIsDamage>unique status</style> you have.",
                 "Illness",
                 ItemTier.Tier3,
                 UltAssets.ViralSmogSprite,